Further refactoring of App.
Moved option parsing code into App.
pandoc.hs is now a 2-liner.
